requirement.txt
openai
python-dotenv
requests
dotenv
opencv-python
SpeechRecognition
pyaudio
gtts
pydub
pyserial